Factor out the newly-added selection items into a variable
This is necessary for the commit that follows this one, since D3 does
not like this construction…

         statuses.enter().append("circle")
                         .attr("r", self.statusCircleRadius)
                         …
                         .transition().duration(200)
                             …
                 .merge(statuses)
                         .attr("cx", function (d) { return self.xScale(d.x) })
                         …

…because the .transition() call changes the object on which we're
chaining method calls. Instead, we have to capture the result of the
.enter().append() calls (into a variable such as "newStatuses"). Then we
can call newStatuses.transition() and newStatuses.merge() separately, so
that .merge() is called on the result of .enter().append(), not
.transition().

If D3 were to have something like a .pop() or .endTransition() similar
to jQuery's .end(), we could use that instead, but I haven't found such.